+++ app/components/folio/console/form/footer_component.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
 class Folio::Console::Form::FooterComponent < Folio::Console::ApplicationComponent
-  def initialize(f: nil, preview_path: nil, share_preview: false)
+  def initialize(f: nil, preview_path: nil, share_preview: false, show_settings: true)
     @f = f
     @preview_path = preview_path
     @share_preview = share_preview
+    @show_settings = show_settings
   end
 
   def before_render