+++ app/helpers/folio/console/forms_helper.rb
@@ -109,7 +109,9 @@ module Folio::Console::FormsHelper
   end
 
   def form_footer(f, opts = {})
-    if f && f.object && f.object.persisted? && f.object.class.try(:use_preview_tokens?)
+    if opts.key?(:share_preview)
+      share_preview = opts[:share_preview]
+    elsif f && f.object && f.object.persisted? && f.object.class.try(:use_preview_tokens?)
       share_preview = true
 
       content_for(:modals) do
@@ -119,8 +121,11 @@ module Folio::Console::FormsHelper
       share_preview = false
     end
 
+    show_settings = opts[:show_settings].nil? || opts[:show_settings]
+
     render(Folio::Console::Form::FooterComponent.new(f:,
                                                      preview_path: opts[:preview_path],
-                                                     share_preview:))
+                                                     share_preview:,
+                                                     show_settings:))
   end
 end